Rename timeelapsed -> benchmarking
