+++ src/libs/antares/CMakeLists.txt
@@ -332,18 +332,18 @@ set(SRC_XCAST
 	)
 source_group("study\\xcast" FILES ${SRC_XCAST})
 
-set(SRC_TIMEELAPSED
-	timeelapsed.h
-	timeelapsed/timer.h
-	timeelapsed/timer.cpp
-    timeelapsed/content_handler.h
-    timeelapsed/content_handler.cpp
-    timeelapsed/info.h
-    timeelapsed/i_writer.h
-    timeelapsed/csv_writer.h
-    timeelapsed/csv_writer.cpp
+set(SRC_BENCHMARKING
+	benchmarking.h
+	benchmarking/timer.h
+	benchmarking/timer.cpp
+    benchmarking/content_handler.h
+    benchmarking/content_handler.cpp
+    benchmarking/info.h
+    benchmarking/i_writer.h
+    benchmarking/csv_writer.h
+    benchmarking/csv_writer.cpp
 	)
-source_group("misc\\timeelapsed" FILES ${SRC_TIMEELAPSED})
+source_group("misc\\benchmarking" FILES ${SRC_BENCHMARKING})
 
 set(SRC_IO
 	io/statistics.h
@@ -538,7 +538,7 @@ macro(ADD_ANTARES_CORE name)
 		${SRC_INI}
 		${SRC_IO}
 		${SRC_MATRIX}
-		${SRC_TIMEELAPSED}
+		${SRC_BENCHMARKING}
 		${SRC_AREAS}
 		${SRC_LINKS}
 		${SRC_XCAST}
