+++ src/solver/simulation/sim_algorithme_de_flot.cpp
@@ -31,7 +31,6 @@
 #include "sim_structure_probleme_economique.h"
 #include "sim_structure_probleme_adequation.h"
 #include "sim_extern_variables_globales.h"
-#include <antares/timeelapsed.h>
 
 void SIM_AlgorithmeDeFlot()
 {
